Support time range to query dimension names/values
At present, dimensions are not scoped by time window, which makes
dimension related queries to large databases timeout because it searches
all of time instead of a time window specified on the grafana app.

This commit implements the client side changes to address issue by
scoping the search to time range specified on the app. It also bumps up
the plugin version to 1.3.0 since it is a minor backward compatible API
change.

Depends-On: https://review.opendev.org/#/c/670318
Change-Id: I13614595707697552119926273f06967b8d49b1d
Story: 2006204
Task: 35789